Updated docs and game speeds

+++ src/meta.rs
@@ -2,10 +2,10 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Copy)]
 pub struct MetaData {
-    pub game_speed: f64,
+    pub game_speed: u32,
 }
 impl MetaData {
     pub fn new() -> MetaData {
-        MetaData { game_speed: 100.0 }
+        MetaData { game_speed: 100 }
     }
 }
